pyproject.toml
[project]
name = "gwas-catalogue-explorer"
version = "0.1.0"
description = "Exploring GWAS Catalogue data: associations, traits, and summary statistics"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32.5",
    "pandas>=3.0.1",
    "numpy>=2.4.1",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "scipy>=1.17.1",
    "statsmodels>=0.14.6",
    "jupyter>=1.1.1",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "ruff>=0.11",
    "pyright>=1.1",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "basic"
